Add precision recall skeletal

+++ dataloaders/scitsr_graph_rules.py
@@ -248,10 +248,10 @@ class ScitsrGraphRules(Dataset):
             
             y_row = self.cal_row_label(data_row, tbpos)
             y_col = self.cal_col_label(data_col, tbpos)
-            # y_adj = self.cal_adj_label(data, tbpos)
+            # y_adj = self.cal_adj_label(data_row, data_col, y_row, y_col)
 
-            data_row.y = torch.FloatTensor(y_row)
-            data_col.y = torch.FloatTensor(y_col)
+            data_row.y = torch.LongTensor(y_row)
+            data_col.y = torch.LongTensor(y_col)
             data_row.img = img
             data_col.img = img
